requirements.txt
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.4
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.6
django-apscheduler==0.6.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
scout-apm==2.26.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
